Split components in App into separate files
